Feat: Add extension
